+++ src/framing.c
@@ -23,7 +23,7 @@
 #include <stdlib.h>
 #include <limits.h>
 #include <string.h>
-#include <ogg/ogg.h>
+#include "../include/ogg/ogg.h"
 
 /* A complete description of Ogg framing exists in docs/framing.html */
 
@@ -390,7 +390,7 @@ int ogg_stream_packetin(ogg_stream_state *os,ogg_packet *op){
 static int ogg_stream_flush_i(ogg_stream_state *os,ogg_page *og, int force, int nfill){
   int i;
   int vals=0;
-  int maxvals=(os->lacing_fill>255?255:os->lacing_fill);
+  int maxvals=(os->lacing_fill>255?255:(int)(os->lacing_fill));
   int bytes=0;
   long acc=0;
   ogg_int64_t granule_pos=-1;
@@ -650,7 +650,7 @@ char *ogg_sync_buffer(ogg_sync_state *oy, long size){
       return NULL;
     }
     oy->data=ret;
-    oy->storage=newsize;
+    oy->storage=(int)(newsize);
   }
 
   /* expose a segment at least as large as requested at the fill mark */
@@ -860,7 +860,7 @@ int ogg_stream_pagein(ogg_stream_state *os, ogg_page *og){
     int i;
 
     /* unroll previous partial packet (if any) */
-    for(i=os->lacing_packet;i<os->lacing_fill;i++)
+    for(i=(int)(os->lacing_packet);i<os->lacing_fill;i++)
       os->body_fill-=os->lacing_vals[i]&0xff;
     os->lacing_fill=os->lacing_packet;
 
@@ -907,7 +907,7 @@ int ogg_stream_pagein(ogg_stream_state *os, ogg_page *og){
         bos=0;
       }
 
-      if(val<255)saved=os->lacing_fill;
+      if(val<255)saved=(int)(os->lacing_fill);
 
       os->lacing_fill++;
       segptr++;
@@ -979,7 +979,7 @@ static int _packetout(ogg_stream_state *os,ogg_packet *op,int adv){
      segments.  Now we need to group them into packets (or return the
      out of sync markers) */
 
-  int ptr=os->lacing_returned;
+  int ptr=(int)(os->lacing_returned);
 
   if(os->lacing_packet<=ptr)return(0);
